Move the Site class into Facades namespace
