+++ packages/framework/src/Framework/Services/BuildService.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Hyde\Framework\Services;
 
 use Closure;
+use Hyde\Facades\Site;
 use Hyde\Foundation\RouteCollection;
 use Hyde\Framework\Actions\StaticPageBuilder;
 use Hyde\Framework\Concerns\InteractsWithDirectories;
 use Hyde\Hyde;
 use Hyde\Support\Models\Route;
-use Hyde\Support\Models\Site;
 use Illuminate\Console\Concerns\InteractsWithIO;
 use Illuminate\Console\OutputStyle;
 use Illuminate\Support\Facades\File;
