+++ packages/framework/tests/Feature/Commands/BuildSearchCommandTest.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature\Commands;
 
+use Hyde\Facades\Site;
 use Hyde\Framework\Features\BuildTasks\PostBuildTasks\GenerateSearch;
 use Hyde\Hyde;
 use Hyde\Pages\DocumentationPage;
-use Hyde\Support\Models\Site;
 use Hyde\Testing\TestCase;
 
 /**
